requirements.txt

requests>=2.31.0
python-dotenv>=1.0.0
gspread>=5.10.0
google-auth>=2.22.0
pandas>=2.0.0
pybit>=5.6.0